backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
supabase==2.5.1
bcrypt==4.1.2
gunicorn==22.0.0
